Make use of State-provided query data
So far the TabBar has just created a fake Query object representing all
the tasks known to the program. With this change we move the creation of
this fake query into the State object. The TabBar, in turn, can now
simply take care of visualizing all queries the State object contains.