Make TabBar widget support selection of different tabs
The TabBar widget, although envisioned to manage multiple tabs from the
start, never actually got the capabilities to switch between the
different tabs, as we so far had no way to populate them. With this
change we add the necessary logic to allow the widget to select a
different tab, taking care of hiding the previously selected one and
focusing the new one.