Introduce a second configuration file
So far we were concerned solely with managing data directly related to
tasks: a list of tag templates and a list of tasks with associated tags.
In the future we need to store state that is specific to the program
(and potentially dependent on the tasks/tags), but ultimately not
essential. This state, if not present, can always be replaced by sane
defaults.
To accommodate these scenarios, this change introduces a second
configuration file that will store this non-vital data as well as the
infrastructure to load state from it and save it back.