Enable LTO and more optimizations for release builds
The program's binary is almost 1 MiB in size. Although in general that
is not a size to worry about, keeping a small binary and memory
footprint is beneficial in the majority of cases and leaves a
tangentially better impression with users.
To that end, this change enables the following optimizations to be
performed when creating a release build:
1) We compile with optimization for code size. We have no performance
   sensitive code and the program is idling most of the time, meaning
   that binary size and memory size will ultimately have the most weight
   when judging the program. Hence, minimizing those seems like the best
   trade-off.
2) We enable link-time optimization (LTO). At the expense of compilation
   time (which is not a concern for what is generally a one-off
   operation), this step can reduce binary size by eliminating more
   unused code as well as enable performance related optimizations not
   possible without this setting. For similar reasons we disable
   incremental builds and treat the entire compilation as one unit.
3) We abort on panics. This setting cuts down binary size by a not
   insignificant percentage. Because of the "alternate screen" we use we
   currently do not see any panic messages. Thus, removing backtraces
   and other helpful information makes no difference until we have
   solved this problem.
The end result of these optimizations is a reduction of binary size by
more than a fourth (250 KiB).