Remove need for out-of-band accesses to Templates object
With the introduction of the make_tasks_with_tags functionality for
testing we had to also introduce a way to retrieve the Templates object
from the Tasks object in order to instantiate tags. With the move of
ownership of Templates out of Tasks, however, we no longer require this
out-of-band access. Instead, we can directly return the Templates object
to use and simplify the code accordingly. This patch accomplishes this
task.