Iterate over keys instead of all events
As it turns out the termion crate contains functionality to only read
keys as opposed to generic events, which includes mouse input as well.
So far we have simply ignored events we are not interested in, but
clearly the better way is to not receive those events in the first
place.
To that end, this change adjusts the logic to only wait for key events
and no others.