Handle events in batches
We have found that pasting larger amounts of text into the program
causes a noticeable delay. The reason for this behavior is that each
character in the pasted text is input as a key, which is converted into
an event, which will be handled by the UI, which will cause an update
event to be emitted, which in turn will trigger a redraw. The problem
with this flow of events is that redrawing operations are by far the
slowest in the chain and so performing them for each key can make the
input appear sluggish.
In order to circumvent this problem we would like to batch up all
outstanding key inputs, handle them in one go, and then -- if any of
them required a redraw -- perform a render operation. Unfortunately,
neither the termion crate nor Rust in general provide a way to check for
input via a poll(2) style API and so we have no way to know whether a
key is available or not (and, hence, whether we will be blocked or not
waiting for such a key).
To solve this problem we instantiate a background thread whose only task
it is to wait for incoming events and put them into a channel. The
channel is buffered (with an unlimited buffer size) and will be drained
in batches by the main thread. Only after the batch is handled will a
render operation be triggered, should any event have indicated the need
for it.