Refactor task related testing code
The testing code has become a bit of a challenge to work with. In
particular, we failed to establish a proper layering around the
functionality we have. We also have a test-only struct, TaskVec, that is
required for certain task related equality comparisons but used in more
contexts than that.
This change reworks most of this code. The main accomplishment is a
switch from the test-only TaskVec struct to basing everything off of
SerTask object collections. The idea is that SerTask offers the same
comparison functionality that we had to explicitly built for the TaskVec
struct. Furthermore, we do away with the make_tasks_vec function, which
effectively replaced the former make_tasks. Everything is now based on
vectors of SerTask objects.
On the TermUi side of things we introduced a greater variety of
functions instantiating a UI and issuing events, but with different
return values.