Factor out test module
We have minimized but not eliminated the amount of sharing between
individual test sites in the code. The remaining shared functionality is
either used in many places of the program for good reason or anticipated
to be used that way.
In order to have clearer responsibilities and to form a sort of
hierarchy (which generally make it more intuitive to reason about code),
this change factors out a new module that contains the remaining shared
test code.