Inherit tags of the currently selected task for new tasks
We currently lack the means to configure the tags for a given task. That
is especially problematic for newly added tasks, which rarely have no
tags associated with them.
To work around this problem until a proper UI for tag adjustments is
implemented, this patch adds logic to inherit the tags of the currently
selected task. In practice this heuristic was found to yield very
accurate results, but this depends largely on the usage patterns of the
program.