Search all tabs for the task to select
When a new task is added it is supposed to be selected. However, it is
not quite clear which task list box (representing the various queries)
actually displays the task. It is most likely the currently active one,
but it may not be if the task was added in under the view of a query
that only shows completed tasks. In such a case the newly added task
(which is not tagged as complete) will not appear in the query and,
hence, not be visible in the task list box. To provide the best possible
user experience, we should search all tabs and check whether one
displays said task. If one does, select it and be done.