Add support to serialize and deserialize queries
This change finally adds support for serializing and deserializing Query
objects. This is a big mile stone, as it enables us to finally make use
of the tabbed view we implemented earlier. As it stands, however, that
is only possible by manually configuring queries (each Query is
represented as a tab) in the configuration file.
Query objects are ultimately program state and as such are stored in the
recently introduced configuration file for program specific data.
However, they still require access to the available tags, which
complicates loading quite a bit.