Draw indication whether more tabs are available on the left/right
Users are free to create as many queries as they please. As these
queries are visualized in the form of tabs, however, and there is a
limited amount of screen space available to do so, we may only be able
to show some but not all of them.
To make it more obvious that more tabs are available, this change adds
code to draw a small indication whether more tabs will follow on either
the left and/or the right side in the tab bar.