Move renderer specific data from widgets into TermRenderer struct
When rendering widgets we may have to keep some state that is actually
specific to the rendering process and not used by the widget itself. An
example is the offset we maintain which describes the first task/tab to
render. This offset mainly depends on the amount of screen space
available, and as such is specific to the process of rendering,
encapsulated in our case by the TermRenderer. Conceptually it is wrong
to store this data in the widget itself. Nevertheless, we did so in the
past.
This change moves this attribute from the widget (TaskListBox and
TabBar, respectively) into the TermRenderer. Because these attributes
need to be maintained on a per-widget basis, we use the widget ID to map
to the corresponding data object.