Use a shared Tasks object for task related operations
In the past we have designed the program in such a way that the Tasks
object managing all available tasks was shared only between the State
(which eventually will want to persist those tasks) and the TermUi (some
of its widgets may need to display tasks and/or modify them). What ended
up happening was that in order to add a new task, or update or remove an
existing one, an event needed to be sent from the widget requesting the
operation to the TermUi. While conceptually fine, sending of events
involves a lot of boilerplate and is not as easy to understand or debug
as would be ideal.
With this patch we redesign this architecture slightly. Instead of
sharing this object only between the State and the TermUi we also share
it between all widgets that have a reason to require access to it. This
refactoring removes a bunch of events that are no longer needed, while
keeping the majority of the infrastructure the way it is.

This change is part of a rethink of the internal architecture evolving
around the sharing of state. In a nutshell, what we have are a bunch of
widgets that modify state. Eventually this state needs to be persisted
in a file. Two approaches have been identified to make this happen:
1) We can share data at the level of a widget. That means that there is
   no global state object and each widget owns the data it manipulates
   (to the degree possible; tasks still need to be shared among many
   widgets, but sharing could happen at the tab bar level as opposed to
   with an entity completely outside of the UI -- the State).
2) We keep references to the all the data we want to persist in a global
   object and hand out references to the individual pieces to the
   requesters on demand.

The first is conceptually more complicated and diverges from the
existing architecture more greatly. The problem is that in order to save
the entire state into a file it would first need to be collected. That
would imply sending events to all widgets and then assembling their
responses into the final object to save. For these reasons we are moving
the program more in a direction resembling the second direction.