Add support for reordering tasks
A frequently wished for operation on tasks is to reorder them.
Reordering in the user's sense refers to changing the order relative to
another task, typically by swapping two adjacent tasks repeatedly.
Because of our concept of views, however, what appears to be a simple
swap operation may actually end up being a search for both tasks and the
reordering of the first one just before or after the other.
With this change we hook these steps up with the UI, the TaskListBox in
particular.