Add support for negated tags to Query
At its inception it was already clear that at some point the current
tag-based task filtering support in the Query struct is not sufficient
to express all possible ways of filtering via tags. In particular, what
has been missing is the ability to negate tags. E.g., the constraint
"not tagged as complete" cannot be expressed currently.
This patch adds this functionality. It does so by wrapping the concept
of a tag as it is stored in a query in a "literal". This literal
expresses whether or not the tag is positive (i.e., has to exist) or
negative (must not exist). With this change we are able to express all
imaginable constraints in a query.