Bump version to 0.1.0
This change bumps the version of the crate to 0.1.0, which marks the
first official release.
The program should no longer be considered to be in an early development
phase and completely experimental. It can and is already successfully be
used for task management, although some functionality still requires
manual adjustment of the appropriate configuration file.
This change also updates the README file to reflect the current status
more appropriately.