Refactor main event handling logic
This change refactors the main event handling a bit. That refactoring is
in preparation for an upcoming change that will require us to handle
more than just key events. To that end we introduce a new (private)
event type that will be used to capture all possible kinds of
notifications in the main part. As we ultimately want to register more
than one event source, we also move the creation of the event channel
outside of the receive_keys function.
Note that we introduce yet another internally used event type. It would
conceptually be possible to just deal with gui::Event objects, however,
since they contain Box<Any>, which does not implement Send, we cannot
really send those over thread boundaries.