Handle SIGWINCH to trigger re-render operations on window resize
So far we have had one annoyance in the program: when a terminal window
is resized we will not automatically redraw the UI of the program. It is
only upon the next key press that a render operation is likely to happen
(depending on the action the key triggers, if any).
That is not really acceptable and so this change adds logic to detect
resizes of the terminal window and redraws the UI accordingly. We do so
by intercepting a handler for SIGWINCH. As typical for signal handlers,
working with them is a mess. We use the self-pipe trick to wake up a
sleeping thread which will then issue a resize event that will cause the
UI to be re-rendered.