Replace extern "C" declarations with imports from libc crate
When initially adding the temporary file functionality we wanted to
limit the excessive usage of crates we depend on. For that reason we
declared the testing-only extern "C" functions we required manually,
instead of importing them from the libc crate.
With a recent change we decided to pull in said crate as a full
dependency and, hence, this change replaces the manual function
declarations with imports from libc.