Introduce explicit state tracking in TaskListBox widget
With an upcoming change we will introduce one more client making use of
the InOutArea widget for inputting text. This client, the TabBar,
however, is never actually focused, which means that our approach of
sending an event indicating the text entered directly to the previously
focused widget will not work. Because the focusing based approach was
found to be reasonable, we preserve that, but instead make the
TaskListBox forward events that it is not supposed to handle in its
current state to its parent.
We do so by explicitly keeping track of the state the widget is in. This
state is represented as a newly introduced enum that either indicates
the addition of a task or the editing of one. Because it is wrapped by
an Option (defaulting to None) we get the state space we will need in
the future.
This change also fixes a bug when editing a task and canceling this
edit process. Previously, any follow up task addition would actually
result in editing the task that was previously edited (but canceled).