Remove task when the edited summary is empty
When adding a new task but inputting an empty text as the summary, we
simply don't do anything, as there is little point in having an empty
task float around. For editing, however, we had no such special casing
in place.
With this change we check for an empty summary during editing as well,
and, if found, remove the task in question instead of making its summary
empty.