Enable bare-trait-objects lint in preparation for Rust 2018
In a not-too-distant upgrade of Rust bare trait objects were deprecated
or at least a different syntax was introduced that will eventually
supersede the old one. That is, traits object types should be introduced
with the 'dyn' keyword.
We make quite some use of trait objects. With this change we convert all
occurrences to the new syntax. Along with that we enable the
bare-trait-objects lint to catch regressions.