Enable Rust 2018 and future compatible lints
This change enables the rust-2018-compatibility and rust-2018-idioms
lints. These lints subsume the previous enabled bare-trait-objects and
elided-lifetimes-in-paths ones, which consequentially get removed. We
furthermore enable the future-incompatible lint which hopefully catches
other potential problems early on.
Note that we explicitly allow the unreachable-pub lint. The crate at
hand is a binary, which means visibility is of little concern. Without
this explicit allowance dozens of 'pub' keywords would need to be
changed to 'pub(crate)' for no benefit.