Add support for selecting tabs by numbers
This change adds keyboard short cuts for directly selecting tabs by
their index. This approach is likely to be faster for the user compared
to moving the selection left or right one tab at a time.
It furthermore introduces the notion of the previously selected tab,
which can be activated through a key as well.