Rework selection logic
In order to select a task that is displayed in some (unknown)
TaskListBox we employ some event driven logic. So far this functionality
was implemented in a fairly ad-hoc fashion: we would remember the widget
where we started (i.e., the first widget we checked and that turned out
not to contain the task of interest) and ask subsequent widgets (as
known to the TabBar) to select the task or continue. We would terminate
the algorithm should we reach the first widget again without having
found the task.
Although doing its job just fine, this functionality is not very
accommodating to future needs due to its inflexibility. In particular,
we can only select a task by ID, we terminate after finding it, and we
will never cycle around.
In the future we would like to implement a search feature, based on a
string match on a task's summary, based on this same functionality.
Contrary to what we currently have, the search would be interactive and
we would potentially want to cycle around.
In order to prepare for this change refactors this selection
functionality to be more flexible, reusable, and actually testable. More
concretely, it introduces a Selection enum capturing the various states
a selection can be in as well as a new struct, SelectionState,
representing the state that is sent between the widgets that is required
to correctly handle such selection scenarios.