Add support for searching tasks by summary sub-strings
A quite essential feature of any moderately useful task management
application is the ability to search for already present tasks. So far
we have not supported this use case, but with this change we introduce
functionality to do so.
We use the in/out area to enter the search string. Matches are then
found through sub-string comparisons on a task's summary. All
comparisons happen in a case-insensitive manner.