Remove unnecessary Quit events from tests
We have a bunch of UI related tests that issue a quit event in the form
of a press of the 'q' key. Such an action is not necessary to terminate
the event handling on the UI used for testing, as the set of events
issued is finite and handling will stop automatically at exhaustion. As
such, these events are unnecessary and are removed.