Retrire test_for_tasks and test_for_ser_tasks functionality
This change is the first in a series to refactor the UI testing logic.
The problem we have with our current testing logic is that it cannot be
used in a flexible manner. It is not possible to issue a second set of
events to a UI, for example, as that UI is available only within a
function.
The design we have conceived to cope with this problem is comprised of a
two-struct architecture:
1) We have a builder object that defines how the UI is built. Right now
   we are only concerned with creation from a set of SerTask objects,
   but we are not limited to that.
2) The builder will create a actual UI instance. We can issue events to
   this instance repeatedly. We also have the option to retrieve the
   current set of tasks in different formats and after each stage.

Overall this design is much more compositional and it will enable more
comprehensive testing in the future.
With this change in particular we replace the test_for_ser_tasks and
test_for_tasks functions with this new infrastructure. The remaining
test functions will follow shortly.