Replace std::cell::RefCell with cell::RefCell
This change replaces all usages of std::cell::RefCell with references to
the same name in the cell crate.
This replacement marks a first step in a rework of the Query's API
surface. Currently the interface provided by it is very much a custom
one: although it is designed to be as close as possible to that of an
Iterator, we actually have internal iteration as opposed to external one
-- a major difference. With a future change we plan to move to an
external iteration model, and having a custom RefCell is a necessity for
that, because exposure of an object holding a reference to a borrowed
entity is not possible with std::cell::Ref.