Rework Query interface
This change reworks the Query's API surface. We switch from a model of
internal iteration, i.e., one where a closure is passed in to the
iteration function of interest, to one with external iteration, which is
much more common in the Rust world.
While having "unsurprising" programming interfaces is a good reason in
and of itself, it is not the main driver for this change: we have
problems expressing all the various ways (and combinations thereof) of
iterating over a list of tasks using the internal iterator model. By
exposing an std::iter::Iterator compliant iterator, we actually get all
of those iteration variations for free.
Such an exposure, however, is more complicated than is obvious at first
glance. Because Query objects borrow the underlying list of tasks on
demand by means of a RefCell, we have to make do with RefCell's (and
Ref's) APIs. Unfortunately, those are insufficient for the task at hand.
Ref::map, in particular, is only able to map to other Ref objects, and
those have the problem of containing a true reference. Iterators,
however, actually are values.
To solve this problem the custom cell crate is used, which provides a
map_val function that solves this problem.

That's not quite the end of the story, though, as even usage of map_val
is not quite as straight forward from our intended context. As the
produced object embeds a value, this value's size needs to be known.
That in turn is a problem with std::iter::Filter that we have relied on
so far (by making use of std::iter::Iterator::filter), because while it
stores the predicate it uses as a member we effectively need a closure
to capture over at least parts of a Query object to know what to filter
for (as opposed to just using a function pointer). Closures however,
have an undefined size and type, so they cannot be contained in the
return type and embedded in an object directly (boxing is simply not an
option here).
To that end, we had to introduce a custom Filter struct that stores a
reference to the literals we filter by directly and can filter by them
when used as an iterator.