Do not create an "all" query unconditionally
In the past we have always created a query object capturing all tasks as
the very first query. This query was not stored and its creation could
not be suppressed. This behavior was always meant to be changed
eventually, preferably after queries can be created, configured, and
destroyed through the UI. We are not quite there yet, but we still want
to rework this logic.
With this patch we only create such a query if the user did not
configure any other queries on his/her own. We also store the query. If
that is undesired behavior it can always be removed by manually editing
the config file, which is how queries need to be created to begin with
right now.