Use ExactSizeIterator where possible
We have a couple of locations that require the size of a collection
represented by an iterator. In order to retrieve this size we have
employed the Iterator::count() method. However, using this method has
the downside that it actually consumes the iterator, which is both
expensive (a lot of "next" calls) and also inconvenient (if the iterator
is to be used further it needs to be cloned first).
Certain iterator, however, implement the ExactSizeIterator trait and
with it provide a len() method that reports the size of the iterable
without consuming the iterator. With this change we adjust the sites
that actually have an ExactSizeIterator at hand to make use of this
method.