Rework selection logic for TabBar widget
Similar to the TaskListBox which was adjusted some time ago, the
TabBar's selection logic assumes that the index of the currently
selected tab can never be invalidated. While that invariant is valid as
it stands, it will no longer hold once tabs can be added and removed
freely.
Since such logic is planned in the future, and because there is value in
having similar selection logic in the two widgets, this change reworks
the logic in the TabBar widget to sanitize the indices before each and
every use.