Add support for writing output to a file other than stdout
This change adds support for writing the program's output to a file
other than stdout, which was -- and still is -- the default. The file to
write to can be supplied as path as the first argument to the program.
Having the option to redirect the output to a different file can be
tremendously useful for debugging purposes. When having two terminals,
for example, one can be used for working with the debugger while the
other simply displays the program's output:
  # Terminal 1:
  $ tty
  > /dev/pts/11

  # Terminal 2:
  $ rust-gdb notnow /dev/pts/11

Note that similar functionality could already be accomplished by using
the redirection functionality provided by common shells. However,
invocation with the debugger becomes cumbersome when it is the shell
that is performing the redirection (it likely means we need to attach to
the program started by the shell after it has been started). To that
end, the functionality was implemented in the program proper.
Lastly, note that not all output is redirected. All fatal errors
happening will still be emitted to stderr as before.