Fix stack overflow when a string is not found
When a string is search across multiple tabs and not found, we run into
an endless loop. This problem is due to an incorrect reset of the
selection's "cycle" state in the case where we are just forwarding the
search event to the next tab.
This change fixes the problem and adds a regression test.