Simplify selection logic
In order to keep track of which widget to select next for an operation
such as a search, we use some selection tracking logic. This logic,
however, is unwieldy and complex. It is covered by enough unit tests to
be confident in its workings, but complexity would increase every more
due to anticipated future work. In particular, we need to support
searching in reverse order and the existing logic we have does not lend
itself to an implementation of the same.
Thus, this change reworks the logic. We simplify it by removing the
existing Cycled state (which was used to keep track of when we visited
all widgets once, in order to prevent endless loops) and instead store
the total number of advancements/increments directly.