Correctly search through tasks on the same tab
When a search is started at a certain task on a certain tab, and the
only possible results are to be found on the very same tab but before
the task where the search started, we are unable to find those matching
tasks currently. The problem got introduced with the recent change to
simplify the selection logic. We report to have "cycled" once we reach
the same widget, but that is actually too early. We should only indicate
that after we went one widget further. Ultimately it does not matter too
much when we indicate that we "cycled", as long as we visited all
widgets at least once and stop eventually.
With this change we adjust the selection logic to only report that we
"cycled" through all widgets after we went past the widget we started
the search at.