Declare selection advancements as signed
With the desire to implement the capability to search for a string in
reverse order, we need to be able to advance by a negative amount. As a
starting point, that also means that the number of advancements needs to
be stored in a signed type. This change adjusts the declaration
accordingly.