Implement support for searching in reverse order
Although we have supported searching tasks by sub-strings contained in
their summary for a while now, that search was only possible in
"forward" direction. Sometimes, however, it is more appropriate to start
a reverse search (similar to the default Vim binding of '?') or to
alternate the search direction mid-way ('N').
This change adds support for searching in reverse order. Note that we do
not exactly implement the same behavior as Vim. In particular, the
"reverse" state when starting a search with '?' applies only to the
first result. A subsequent 'n' would switch back in normal order while
'N' would continue in reverse.