Factor out ToSerde trait
With an upcoming change we plan to move the Id functionality into a
crate of its own to foster code reuse. The functionality for converting
objects into a serializable form, however, is of no interest for the
generic version.
Because we still want to add such additional functionality on top,
however, this change introduces a custom trait that allows us to do just
that. In addition, factoring out a dedicated trait for such
functionality is generally a nice-to-have anyway, as it makes
commonalities more obvious by looking at the code.