Unify Return and Esc handling paths in InOutArea widget
We have different handlers for the Key::Esc and Key::Return key events
in the InOutArea widget. Most of the logic, however, is the same. This
change merges the two paths into a single one, making the code a tad
more concise while also eliminating unnecessary divergence in coding
style.