Factor out key handler function
One of the upcoming changes will introduce an alternative way for
processing input in the InOutArea widget that will rely on the
libreadline system library.
In order to make the two alternatives easily replaceable with one
another, this change factors out a new function containing the
pertaining logic for handling keys.