Use change_state method over directly manipulating in_out member
The InOut state in the InOutArea widget is manipulated on a lot of
occasions, simply because it represents, among others, the text being
displayed along with the cursor position. We already have a method doing
this job for us and conditionally emitting an update event if something
truly changed.
With this change we make use of this existing functionality.