Adhere to 2018 edition module system improvements
With the upgrade of the Rust toolchain to 1.30 a couple of module system
related improvements were stabilized. As part of these, we should now
prefix absolute module paths with 'self', 'super', 'crate', or an
external crate name.
This change makes the program adhere to this new set of constraints by
using 'crate::' style use statements.