Rename EventUpdated trait to EventUpdate
Shortly we will introduce a new trait that can be used to check whether
an event has been updated. The existing trait EventUpdated has a name
that would be more apt for this new event than it is for what it
currently provides.
Hence, to be able to use this name in the future, this change renames
the EventUpdated trait to EventUpdate.