Introduce new EventUpdated trait
As part of the effort to reduce the number of redraw operations
happening we introduced functionality to check whether an ChainEvent
contains a TermUiEvent::Updated variant. This functionality was
introduced in the form of a function, is_updated, that can work on an
UiEvents object. In the future we want to provide similar functionality
for UnhandledEvents as well.
To be able to do so, this change factors out a trait, EventUpdated, that
can be implemented for different event types.