Add test checking for expected update events being emitted
This change adds a test that checks that an update event is only
reported for certain actions in the UI. This is not really a targeted
test, and similar in nature to what we already employ for the InOutArea
state. It also just assumes an empty UI and, hence, does not test all
possible state transitions. However, it is better than having nothing at
all.
As part of this change we also fix a bug discovered by the test where
the TaskListBox widget causes a panic when the space key is pressed with
no tasks being present.