Introduce InOutState struct
This change introduces a wrapper, InOutState, around the existing InOut
struct, that is used internally to the InOutArea widget. This wrapper's
purpose is to associate a generation number with the current InOut
state. This generation number is incremented on every change of the
state itself.
This functionality will help us in the future to avoid unnecessary
redraws due to spurious InOut state changes in the widget.