Prevent unnecessary redraws due to InOut state change
We have seen unnecessary redraw operations happening when, for example,
we press the 'n' key repeatedly and no the search does not yield any
results. The reason for this redraw is that we install an event hook in
the InOutArea widget that is triggered on all events (we only support
key events currently). It then takes care of reverting the widget's
state to "Clear". That is desired because we want status texts to be
cleared after any key press.
However, what ends up happening in cases such as the one mentioned above
is that we clear the state inside the hook and then actually set it back
to the previous value. Now, because we had a reset in between, the
program logic infers that an actual state change has occurred and
triggers a redraw.
To resolve this problem we now just emit an event from the event hook.
Said event is sent to our widget, but the gui crate guarantees that this
event will be received only after all other non-hook emitted events are
processed. So when we receive this hook-emitted event we ultimately know
whether the state changed, and only clear it if that was not the case to
prevent spurious redraws.
With this change we also modify the logic when to set the state to
"Clear" itself, from one that does so only when we are not focused to
one that checks for whether we are not in the "Input" state. While both
conditions theoretically should be true under the same circumstances,
the latter appears to be more directly expressing what we want.