Use rline crate for inputting text in InOutArea widget
Despite having our custom rolled functionality for cursor handling when
inputting text (e.g., for moving the cursor, deleting characters,
inserting characters at certain locations), ultimately that may still be
considered a foreign behavior when compared to how input handling can
work on the command line. A library employed by common shells (at least
on Linux) that handles input of text is libreadline. This library
supports among Emacs style text navigation and input shortcuts, a Vim
mode in which typical operations such as "advance the cursor by one
word" are implemented.
With this change we use the rline crate for interfacing with libreadline
and providing the same input experience as we would on the command line
when entering text through the InOutArea widget.
Note that this functionality is guarded by a feature switch, readline,
for two reasons:
1) libreadline is a dependency that may or may not be desired and/or
   available on a system.
2) Because libreadline reads the system configuration and is highly
   customizable, we can't really run all our tests using it. Some of the
   tests work with raw key events and different configurations would
   behave differently. Also, even in the default configuration not all
   existing bindings would yield the same result.