Remove unknown_lints lint from the allowed list
In the past we have disabled warnings for the 'unknown_lints' lint
because of the assumption that only clippy will have such unknown lints.
As it turns out, a compiler upgrade can actually remove previously
existing lints. If that happens, we would want to eventually remove
those.
To that end this change removes the 'unknown_lints` lint from the
allowed list, causing warnings to be emitted for unknown ones. Along
with that change we also remove the 'unnameable_test_functions' lint,
which seemingly does not longer exist with Rust 1.30, and moves clippy
specific lints into a dedicated cfg block.