Add Cargo.toml categories
For a while now Cargo has supported the specification of categories in
the Cargo.toml file and crates.io will actually honor those categories
and show case the crate in them.
With this change we specify the two categories this crate is believed
to fit in the best: 'command-line-utilities' and 'text-editors'.