Ignore multi-byte character input
Our character handling logic, being based on Rust's String struct, is
Unicode capable at its core. However, because we replace parts of said
string and the indexing done as part of that is happening at the byte
level (as opposed to the grapheme level), we may actually invalidate the
string which is represented in UTF-8 encoding by violating character
boundaries.
Unfortunately, this is very problematic to fix. While the String struct
not only supports byte-wise access but is also character aware, a
character in Rust's terms may not necessarily constitute a character as
the user would recognize it. Special characters such as tilde or grave
accent "decorating" a vowel, for example, may be treated as a separate
character.
The only real solution to this problem comes in the form of the
unicode_segmentation crate would actually allow us to honor grapheme
boundaries. However, the effort to convert everything appears to be high
and we would add another dependency for functionality that is rarely
used. As such, we punt on that right now and just ignore all multi-byte
characters all together, and, with that, prevent them from causing
havoc. This may or may not be a temporary solution.