Enable unused-results lint
This change enables the unused-results lint which potentially helps
catch bugs introduced by forgetting the evaluation of results. It
appears unlikely that it will cause problems as Option and Result
objects should already be tagged to require evaluation and those seem to
be the more likely sources of trouble. Regardless, the enablement of
this lint seems to be benign.