Favor Vec::into_iter() method over drain()
In the past we have used the drain() method of the Vec struct. Because
that method does not consume the object of interest, erroneous uses can
occur and conceptually unnecessary mutable bindings are created. As it
turns out the into_iter() method is what we were actually looking for,
in that it consumes the object in question.
This change adjusts the code to prefer the usage of Vec::into_iter()
over Vec::drain().