Remove Result return from TermUi::new()
Upon its inception, the TermUi::new() constructor cleared the display
using a write operation. This operation could fail and, hence, the
constructor could fail as well. To report the error the constructor
returned a Result.
However, as part of a refactoring the clearing was moved into a
different module. Because of that we no longer have any error return
paths. With this change we remove the Result from the return type.